+++ src/app/main-nav/main-nav.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { CalipsoplusService } from '../calipsoplus.service';
 import { Router } from '@angular/router';
 import { environment } from '../../environments/environment';
+import {AuthenticationService} from '../authentication.service';
 
 @Component({
   selector: 'app-main-nav',
@@ -10,52 +11,53 @@ import { environment } from '../../environments/environment';
 })
 export class MainNavComponent implements OnInit {
   constructor(
+    private authService: AuthenticationService,
     private calipsoService: CalipsoplusService,
     private router: Router
   ) {}
 
   ngOnInit() {
     if (this.router.url === '/autologin') {
-      this.calipsoService.unauthUmbrella().subscribe(unauth_resp => {
-        this.calipsoService.getUmbrellaSession().subscribe(
+      this.authService.unauthUmbrella().subscribe(unauth_resp => {
+        this.authService.getUmbrellaSession().subscribe(
           session_resp => {
             const username = session_resp.uid;
             const EAAHash = session_resp.EAAHash;
 
-            this.calipsoService.authWithEAAHash(username, EAAHash).subscribe(
+            this.authService.authWithEAAHash(username, EAAHash).subscribe(
               user_resp => {
                 // console.log(user_resp);
-                this.calipsoService.login(username, 'umbrella');
+                this.authService.login(username, 'umbrella');
                 this.router.navigate(['/login']);
               },
               error => {
                 if (error.status === 404) {
                   // console.log("User in UO not found or not linked..");
-                  this.calipsoService.goExternalLoginWOU();
+                  this.authService.goExternalLoginWOU();
                 }
               }
             );
           },
           error => {
             // console.log("No session found.. go to umbrella page");
-            this.calipsoService.goExternalLoginUmbrella();
+            this.authService.goExternalLoginUmbrella();
           }
         );
       });
     }
   }
   public getUserName(): string {
-    return this.calipsoService.getLoggedUserName();
+    return this.authService.getLoggedUserName();
   }
 
   public isLogged(): boolean {
-    return this.calipsoService.isLogged();
+    return this.authService.isLogged();
   }
 
   public logout() {
     // console.log("login_local:"+this.calipsoService.calipsoSettings.local_auth);
-    if (this.calipsoService.getLoginType() === 'local') {
-      this.calipsoService.unauth().subscribe(
+    if (this.authService.getLoginType() === 'local') {
+      this.authService.unauth().subscribe(
         resp => {
           // console.log("logout done from UO");
           this.router.navigate(['/']);
@@ -65,7 +67,7 @@ export class MainNavComponent implements OnInit {
         }
       );
     } else {
-      this.calipsoService.unauthUmbrella().subscribe(
+      this.authService.unauthUmbrella().subscribe(
         resp => {
           // console.log("logout done from umbrella");
           window.location.href =
