+++ src/app/select-calipso-quota-form/select-calipso-quota-form.component.ts
@@ -6,6 +6,7 @@ import { CalipsoImage } from '../calipso-image';
 import { CalipsoplusService } from '../calipsoplus.service';
 
 import { Router } from '@angular/router';
+import {AuthenticationService} from '../authentication.service';
 
 @Component({
   selector: 'app-select-calipso-quota-form',
@@ -14,6 +15,7 @@ import { Router } from '@angular/router';
 })
 export class SelectCalipsoQuotaFormComponent implements OnInit {
   constructor(
+    private authService: AuthenticationService,
     private calipsoService: CalipsoplusService,
     private router: Router
   ) { }
@@ -24,7 +26,7 @@ export class SelectCalipsoQuotaFormComponent implements OnInit {
   available_quota: CalipsoQuota = new CalipsoQuota(0, 0, '0', '0');
 
   ngOnInit() {
-      const username = this.calipsoService.getLoggedUserName();
+      const username = this.authService.getLoggedUserName();
       this.calipsoService.getCalipsoQuota(username).subscribe(
         quotas => {
           console.log('quotas:' + quotas.cpu, quotas.hdd, quotas.max_simultaneous, quotas.memory);
